% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lisi}
\alias{compute_lisi}
\title{Compute Local Inverse Simpson's Index (LISI)}
\usage{
compute_lisi(X, meta_data, label_colnames, perplexity = 30, nn_eps = 0)
}
\arguments{
\item{X}{sample by feature matrix.}

\item{meta_data}{Dataframe with cell labels.}

\item{label_colnames}{Which variables to compute LISI for.}
}
\value{
Table (data.frame) of LISI values. Each row is a cell and each column is a different label variable.
}
\description{
Use this function to compute LISI scores of one or more labels.
}
\examples{

## Example with 400 cells. 
library(lisi)
head(lisi::meta_data)

## Let's color cells by labels. For label 1, there are mixed and non-mixed groups. 
## for label 2, all cells are well mixed. 
lisi::X \%>\% 
  cbind(lisi::meta_data) \%>\% 
  dplyr::sample_frac(1L, FALSE) \%>\% 
  tidyr::gather(key, val, label1, label2) \%>\% 
  ggplot(aes(X1, X2, color = val)) + geom_point(shape = 21) + 
  facet_wrap(~key)
  
## now to compute and plot the LISI values for each label. 
lisi_res <- lisi::compute_lisi(lisi::X, lisi::meta_data, c('label1', 'label2'))
head(lisi_res)

lisi::X \%>\% 
  cbind(lisi_res) \%>\% 
  dplyr::sample_frac(1L, FALSE) \%>\% 
  tidyr::gather(key, lisi_value, label1, label2) \%>\% 
  ggplot(aes(X1, X2, color = lisi_value)) + geom_point(shape = 21) + 
  facet_wrap(~key)

  
}
