% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_simpson_index}
\alias{compute_simpson_index}
\title{Compute the Local Inverse Simpson Index (LISI)}
\usage{
compute_simpson_index(
  D,
  knn_idx,
  batch_labels,
  n_batches,
  perplexity = 15,
  tol = 1e-05
)
}
\arguments{
\item{D}{Distance matrix of K nearest neighbors.}

\item{knn_idx}{Adjacency matrix of K nearest neighbors.}

\item{batch_labels}{A categorical variable.}

\item{n_batches}{The number of categories in the categorical variable.}

\item{perplexity}{The effective number of neighbors around each cell.}

\item{tol}{Stop when the score converges to this tolerance.}
}
\description{
Compute the Local Inverse Simpson Index (LISI)
}
