% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{Metadata table.}
\format{A data.frame with 400 rows and 2 columns}
\usage{
meta_data
}
\description{
Table containing two types of labels for the simulated cells.
}
\keyword{datasets}
